requirements.txt
# Добавьте сюда ваши зависимости Python
Django==3.2.20
djangorestframework==3.14.0
django-filter==23.1
django-storages==1.12.3
psycopg2-binary==2.9.9
label-studio
label-studio-sdk==0.0.30
label-studio-tools
celery==5.3.6
redis==5.0.1
flower==2.0.1
 Pillow==10.0.1
gunicorn==21.2.0
google-cloud-storage
django-health-check